+++ shared/modules/bus_module.go
@@ -3,9 +3,6 @@ package modules
 //go:generate mockgen -source=$GOFILE -destination=./mocks/bus_module_mock.go -aux_files=github.com/pokt-network/pocket/shared/modules=module.go
 
 import (
-	"encoding/json"
-
-	"github.com/benbjohnson/clock"
 	"github.com/pokt-network/pocket/shared/debug"
 )
 
@@ -27,10 +24,6 @@ type Bus interface {
 	GetConsensusModule() ConsensusModule
 	GetTelemetryModule() TelemetryModule
 
-	// Configuration
-	GetConfig() map[string]json.RawMessage
-	GetGenesis() map[string]json.RawMessage
-
-	// Time
-	GetClock() clock.Clock
+	// Runtime
+	GetRuntimeMgr() RuntimeMgr
 }